move api routes to api.py and add some docs
