+++ README.md
@@ -9,17 +9,16 @@ Features:
 - Everything is a file! For ease of access and editing, all the content is stored in markdown files with yaml front matter.
 - Extensible search with Elasticsearch and its Query DSL
 - Allows you to sync up with Pocket to gather bookmarks from there too.
+- dark theme
 
 
 ![demo (low res)](https://github.com/Uzay-G/archivy/raw/master/archivy.gif)
 
+
 Upcoming:
 
-- Integrations with HN, Reddit, and many more.
-- Login module
-- Add submodules for digital identity so archivy syncs to your hn upvoted posts, reddit saved, etc...
+- Plugin system to allow people to publish and write extensions to archivy
 - Option to compile data to a static site that can be deployed.
-- Dark theme
 - UI for grouping by tag and use NLP to automatically generate connections between posts
 
 ## Setup
@@ -87,6 +86,12 @@ These credentials will be printed to the log when you launch like this:
 Login with these credentials and then you can change your password/username by clicking the profile button on the top left.
 
 You can then use archivy to create notes, organize it and store information.
+
+## Scripting
+
+You might be interested in extending archivy by for example building scripts that allow you to regularly fetch data or other functionalities. In that case you can use the api system and soon a [powerful plugin system](https://github.com/Uzay-G/archivy/issues/86).
+
+
 
 ## Community and Development
