+++ archivy/routes.py
@@ -10,7 +10,6 @@ from werkzeug.security import check_password_hash, generate_password_hash
 from flask_login import login_user, login_required, current_user, logout_user
 
 from archivy.models import DataObj, User
-from archivy.search import query_index
 from archivy import data, app, forms
 from archivy.extensions import get_db
 from archivy.config import Config
@@ -122,31 +121,6 @@ def delete_data(dataobj_id):
     return redirect("/")
 
 
-@app.route("/folders/new", methods=["POST"])
-def create_folder():
-    directory = request.json.get("paths")
-    try:
-        sanitized_name = data.create_dir(directory)
-    except FileExistsError:
-        return "Directory already exists", 401
-    return sanitized_name, 200
-
-
-@app.route("/folders/delete", methods=["DELETE"])
-def delete_folder():
-    directory = request.json.get("name")
-    if directory == "":
-        return "Cannot delete root dir", 401
-    if data.delete_dir(directory):
-        return "Successfully deleted", 200
-    return "Not found", 404
-
-
-@app.route("/search", methods=["GET"])
-def search_elastic():
-    query = request.args.get("query")
-    search_results = query_index(Config.INDEX_NAME, query)
-    return jsonify(search_results)
 
 
 @app.route("/login", methods=["GET", "POST"])
